[36] feat : 치아 선택 컴포넌트 분리
