+++ src/components/medical/MedicalWrite.tsx
@@ -9,6 +9,7 @@ import ClinicInput from "@/components/medical/ClinicInput";
 import DateInput from "@/components/medical/DateInput";
 import TreatmentSelection from "./TreatmentSelection";
 import CostInput from "./CostInput";
+import ToothSelection from "./ToothSelection";
 
 type ToothSide = {
   value: "left" | "right";
@@ -31,16 +32,7 @@ const MedicalWrite = () => {
   const [toothSelect, setToothSelect] = useState<"left" | "right">("left");
   const [isShare, setIsShare] = useState<boolean>(true);
   const [isClinic, setIsClinic] = useState<boolean>(false);
-  const toothSide: ToothSide[] = [
-    {
-      value: "left",
-      name: "왼쪽"
-    },
-    {
-      value: "right",
-      name: "오른쪽"
-    }
-  ];
+
   const shareButton: ShareButton[] = [
     { label: "네", value: true },
     { label: "아니요", value: false }
@@ -60,50 +52,7 @@ const MedicalWrite = () => {
         {clickTreatment && (
           <>
             <CostInput />
-            <motion.div
-              className={styles.writeWrapper}
-              initial={{ opacity: 0, x: 100 }}
-              animate={{ opacity: 1, x: 0 }}
-              exit={{ opacity: 0, x: -100 }}
-              transition={{ duration: 0.5, ease: "easeInOut" }}
-            >
-              <label className={styles.writeLabel}>
-                치아를 선택해서 받은 치료를 <br /> 설정해 주세요.
-              </label>
-              <span className={styles.helperText}>
-                스케일링과 잇몸치료는 설정에서 제외됩니다.
-              </span>
-              <div className={styles.toothItem}>
-                <div className={styles.toothSelect}>
-                  {toothSide.map((side) => (
-                    <button
-                      type="button"
-                      key={side.value}
-                      className={[
-                        styles.treatmentButton,
-                        side.value === toothSelect ? styles.selected : ""
-                      ].join(" ")}
-                      onClick={() => setToothSelect(side.value)}
-                    >
-                      {side.name}
-                    </button>
-                  ))}
-                </div>
-                <div className={styles.toothImg}>
-                  <AnimatePresence mode="wait">
-                    <motion.div
-                      key={toothSelect}
-                      initial={{ opacity: 0, x: 100 }}
-                      animate={{ opacity: 1, x: 0 }}
-                      exit={{ opacity: 0, x: -100 }}
-                      transition={{ duration: 0.5 }}
-                    >
-                      <Tooth location={toothSelect} />
-                    </motion.div>
-                  </AnimatePresence>
-                </div>
-              </div>
-            </motion.div>
+            <ToothSelection />
             <motion.div
               className={styles.writeWrapper}
               initial={{ opacity: 0, x: 100 }}
