+++ src/components/medical/ToothSelection.module.scss
@@ -1,0 +1,39 @@
+@import "@/styles/common.scss";
+.writeWrapper {
+  position: relative;
+}
+
+.writeLabel {
+  word-break: auto-phrase;
+  @include font-style($semi-subtitle1);
+}
+.helperText {
+  display: block;
+  @include font-style($body2);
+  color: $grey3;
+  margin-top: 5px;
+}
+.toothItem {
+  margin-top: 7px;
+}
+.toothSelect {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+}
+
+.treatmentButton {
+  border: 1px solid $grey1;
+  border-radius: 8px;
+  padding: 7px 16px;
+  @include font-style($body2);
+  color: $black;
+  cursor: pointer;
+}
+
+.selected {
+  @include button-select;
+}
+.toothImg {
+  margin-top: 12px;
+}
